+++ Sources/QRCodeReaderViewControllerBuilder.swift
@@ -30,14 +30,14 @@ import Foundation
  The QRCodeViewControllerBuilder aims to create a simple configuration object for
  the QRCode view controller.
  */
-public final class QRCodeViewControllerBuilder {
+public final class QRCodeReaderViewControllerBuilder {
   // MARK: - Configuring the QRCodeViewController Objects
 
   /**
    The builder block.
    The block gives a reference of builder you can configure.
    */
-  public typealias QRCodeViewControllerBuilderBlock = (builder: QRCodeViewControllerBuilder) -> Void
+  public typealias QRCodeReaderViewControllerBuilderBlock = (QRCodeReaderViewControllerBuilder) -> Void
 
   /**
    The title to use for the cancel button.
@@ -53,6 +53,11 @@ public final class QRCodeViewControllerBuilder {
    Flag to know whether the view controller start scanning the codes when the view will appear.
    */
   public var startScanningAtLoad = true
+
+  /**
+   Flag to display the cancel button.
+   */
+  public var showCancelButton = true
 
   /**
    Flag to display the switch camera button.
@@ -77,11 +82,11 @@ public final class QRCodeViewControllerBuilder {
   public init() {}
 
   /**
-   Initialize a QRCodeViewController builder with default values.
+   Initialize a QRCodeReaderViewController builder with default values.
 
-   - parameter buildBlock: A QRCodeViewController builder block to configure itself.
+   - parameter buildBlock: A QRCodeReaderViewController builder block to configure itself.
    */
-  public init(@noescape buildBlock: QRCodeViewControllerBuilderBlock) {
-    buildBlock(builder: self)
+  public init(buildBlock: QRCodeReaderViewControllerBuilderBlock) {
+    buildBlock(self)
   }
 }